pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "halocli"
version = "0.3.2"
description = "Standalone HaloPSA CLI for safe operator and automation workflows"
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.10"
authors = [{name = "Midtown Technology Group", email = "info@midtowntg.com"}]
dependencies = [
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "platformdirs>=4.0.0",
    "pyyaml>=6.0.0",
    "keyring>=25.0.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]

[project.optional-dependencies]
dev = [
    "build>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[project.scripts]
halocli = "halocli.cli:main"

[project.urls]
Homepage = "https://github.com/Midtown-Technology-Group/halocli"
Repository = "https://github.com/Midtown-Technology-Group/halocli"
Issues = "https://github.com/Midtown-Technology-Group/halocli/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"
